+++ CHANGELOG.md
@@ -3,6 +3,7 @@
 ## master
 
 * Added `is_in_intro_offer_period` to `ReceiptVerifier.IAPReceipt`.
+* Added `:auto` to the `env` option which make it explicitly to retry in different sandbox and production env.
 
 ## v0.9.0